Cyrasterize recipe
This is a work in progress, it builds successfully but I haven't
checked if it actually generates correct results/links properly.

It also relies on a branch of cyrasterize that changes the C
nature of the project to C++. It really needs to be tested on
Linux and Mac as I made a whole load of changes.